Removed the fetch from workouts and put in ProgramList
